+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 5.0.1 (pending)
 
 * properly report server name if it returns eof
+* added `ENGINE_LOG` server configuration
 
 
 ### 5.0.0 (2019-06-24)